pyproject.toml
[project]
name = "skellymodels"
version = "0.1.0"
description = "Mathematically rigorous skeleton, rigid body, and trajectory definitions for motion capture pipelines"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "pydantic>=2.0",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "beartype>=0.18",
]

[project.optional-dependencies]
bvh = [
    "scipy>=1.11",
]
dev = [
    "pytest>=8.0",
]
